Implement preference for mail handler
To avoid configuring a default mail handler for all
Servoshell users, this now becomes a preference.
The preference can be set on the command line to
specify a custom value:

```
./mach run tests/wpt/mozilla/tests/mozilla/resources/mailto.html --mailto-handler="<handler>"
```

It is also shown on the preferences.html page, which
shows an error if the query parameter is set. That
query parameter is set by default as a handler, so that
the user gets shown the error that no handler was configured.

Unfortunately updating the handler in the preferences
doesn't have any effect, since protocols can only be
registered on startup. Therefore, if we want to support
updating the preferences value, we would need to persist
the Servo preferences.

All in all, this allows me to continue testing the
protocol handlers, while also showing embedders how
they can hook up such handlers.

Part of #40615

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>