webgl: Unconditionally decrement busy count of WebGL contexts. (#41094)
The current logic returns early without decrementing the busy count if
the swap chain doesn't have a front surface. This happens if the context
was created but never used for rendering, as seen in the linked issue.
Since the count is always incremented when locking, this causes the
count to never reach zero, causing the cleanup logic in
`remove_webgl_context` to skip freeing the context and associated
resources.

Fixes #41082.

Testing: A new crash test is added.

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>