html: Decouple ownership of the media element from IPC router callback
Added new invokation struct `HTMLMediaElementEventHandler` that allows
to decouple ownership of the `HTMLMediaElement` from the IPC router
callback (break cyclic referrence dependency) by replacing a `Trusted`
wrapper with more suitable `WeakRef`.
Note that `WeakRef` is not thread-safe and MUST be rooted and/or deleted
on the script thread.

Testing: Regressions in the following tests caused by overly aggressive GC
on document activity change to `Inactive` and will be follow up in
https://github.com/servo/servo/issues/41138
- html/semantics/embedded-content/media-elements/autoplay-allowed-by-feature-policy.https.sub.html
- html/semantics/embedded-content/media-elements/autoplay-default-feature-policy.https.sub.html
- html/semantics/embedded-content/media-elements/autoplay-disabled-by-feature-policy.https.sub.html

Fixes: https://github.com/servo/servo/issues/40243

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>