libservo: Remove `WebViewInner::rect`
With the removal of the MDI support, a `WebView` is now always as big as
the `RenderingContext` onto which it renders. This change removes the
remnant of the MDI support which is the `rect` property. This was
problematic because it could get out of sync with the size of the
`RenderingContext`.

This change also adds a test for the case where these could get out of
sync.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>