constellation: Only set up a single Devtools route for all `ScriptThread`s (#41161)
Previously every `ScriptThread` had its own `DevTools` route, but we can
share the same route amongst all of them. This should theoretically save
file handles. The fact that this creating IPC routes is hidden a bit by
the use of `GenericCallback`, but underneath it does actually create a
route.

Testing: This should not show any behavior differences and we do not
test
for the number of file handles used.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>