devtools: Update to Firefox 145 (#41087)
Update the target spec version from 133 to 145. Solves a few issues with
incorrect or outdated messages, such as an error with `reflow`.

The only meaningful change I could find is that the performance actor is
now called "perf" instead of "performance". It is part of a set of
global actors that the root handles. I made it more explicit and moved
the global actor registration inside of the root actor, it makes more
sense to have it there.

Finally, I ordered the message handling alphabetically. The only changed
entry is `getRoot` because of the global actors.

Testing: I manually tested the functionalities and checked the logs.

Signed-off-by: eri <eri@igalia.com>