script: Update `media` to include fix for `AudioBufferSourceNode.loop` (#41075)
Fix in https://github.com/servo/media/pull/476:

Previously:
the moment we set bufferSource.loop = true, the node hit the early
“extreme edge case” check, decided “I refuse to output”, triggered
onended, and returned silence. So the loop never got a chance to run.

Now:

Early bailout is removed(no idea why was there in first place).
The per tick offset is guaranteed to be a sensible value with respect to
the loop range (wrapped if extreme).
The normal mixing code runs, updating pos each tick and respecting the
loop region

Testing: I tested this manually, it would be nice to add a test for it,
and I’m surprised there isn’t one yet.
Fixes: #41073

---------

Signed-off-by: Taym Haddadi <haddadi.taym@gmail.com>