layout: Merge BaseFragment and BaseFragmentInfo
These structs used to be different when they were added in #29699:
`BaseFragment` had a `debug_id` field, while `BaseFragmentInfo` didn't.
But this field was later removed in #35001, so the structs became
identical.

Therefore, this just unified them. I'm picking `BaseFragmentInfo` as the
name, since it was the most frequently used of the two.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>