Fix all Android build warnings and make the CA verifier preference more flexible
This change fixes a variety of build warnings on Android. In a related
change, the CA verifier preference allows using a list of certificate
overrides and the `rust-webpki` verifier regardless of platform. Whether
to use a list of certificate overrides and whether to use the platform
verifier are orthagonal, so this change unlinks those two and in the process
fixes a build warning on Android.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>