servoshell: Use `WheelMode::DeltaPixel` on line-based wheel events converted to pixels
servoshell is converting line-based wheel scrolls to pixels, so we need
to send `WheelMode::DeltaPixel` in with the value. This value is
processed directly by script and is the way it interprets the delta.

A better fix here would be to actually process these values fully in
script, but that is aved for a followup.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>