servoshell: Use `WheelMode::DeltaPixel` on line-based wheel events converted to pixels (#41184)
servoshell is converting line-based wheel scrolls to pixels, so we need
to send `WheelMode::DeltaPixel` in with the value. This value is
processed directly by script and is the way it interprets the delta.

A better fix here would be to actually process these values fully in
script, but that is aved for a followup.

Testing: This part of servoshell is untested, so it's difficult to add
a test for this.
Fixes: #41164.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>