script: Implement storing deferred fetches
This is the first part of preparation for implementing
the keepalive flag for requests. It requires implementation
of the concept of a task group, which we already sort of
have with `DocumentLoader`. Therefore, we currently already
cancel any loads, which should be updated to ignored
kept-alive requests.

However, termination also requires storing of deferred
fetches, which are also kept-alive. Therefore, to distinguish
the two, we need to store them separately as stated with
the TODO.

In a follow-up, we can then skip canceling keep-alive requests
for both `navigator.sendBeacon` (which aren't deferred)
and `fetchLater` (which are deferred).

Part of #41230

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>