fonts: Calculate font keys count via number of font keys and not number of `Painter`s
The changes in #39140 introduced a pretty subtle bug in the way that the
number of font keys to generate are calculated. The code was previously
trying to handle the case where only a single kind of font key needed to
be generated. In that case, one of the stores might not be empty and it
would just generate the difference between the remaining keys and the
max.

In #39140, the data structures were changed to hold a map of `Painter`s
to key stores, yet the calculation lines remained the same. This means
that the number of `Painter`s was subtracted from the maximum number of
keys. This could lead to situations where the number of keys requested
was always 0.

This change fixes that.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>