layout: Do not set `cursor: text` on `<input type=color>` (#41239)
type=color was just missing in the `servo.css`. Since i am new to the
project it would be awesome to get more infos about the testing. Its a
bit overwhelming with all the directories so i am not sure if the test
is okay like this or even in the right place.

Fixes: #41101

---------

Signed-off-by: helvete003 <d.obieglo@gmail.com>