script: Omit GC for `iframe` "about:blank" document with inactive state
If activity state of the document is change to `Inactive`
the user agent is calling garbage collection for the
script global `JSContext` unconditionally and some short-lived
DOM objects (which is not attached to DOM document) is immediately gone.

Let's omit GC for `iframe` document with inactive state and
"about:blank" URL to make GC it less aggressive for top level DOM objects.

Testing: Improvements in the following tests
- /html/semantics/embedded-content/media-elements/autoplay-allowed-by-feature-policy.https.sub.html
- /html/semantics/embedded-content/media-elements/autoplay-default-feature-policy.https.sub.html
- /html/semantics/embedded-content/media-elements/autoplay-disabled-by-feature-policy.https.sub.html

Fixes: https://github.com/servo/servo/issues/41138
Fixes: https://github.com/servo/servo/issues/41193
Fixes: https://github.com/servo/servo/issues/41221

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>