layout: Treat `<audio>` as replaced and do not display when there is no `controls` attribute (#41262)
Add an important UA style to ensure that `<audio>` isn't displayed at
all when it doesn't have the `controls` attribute. This matches Gecko,
Blink and WebKit.

When it does have the attribute, then treat it as replaced. For example,
this means that it won't stretch by default if it's block-level, that it
will be atomic if inline-level, and that it won't generate boxes for its
children (including `::before` and `::after`).

Previously we were generating `::before` and `::after`, but there was a
bug: if they had some replaced content, the controls of the `<audio>`
would be duplicated, and also appear on the pseudo-element. Then, since
the various boxes for the controls would be backed by the same elements,
Servo could panic because of a double borrow. This is now avoided.

Testing: Some WPT are now passing, also adding a new crashtest. There is
one failure about `innerText`, but the test is probably wrong, now we
match other browsers.
Fixes: #40693
Fixes: #41183

Signed-off-by: Oriol Brufau <obrufau@igalia.com>