layout: Remove centralized `FontKeyAndMetrics` list from the `InlineFormattingContext`  (#41208)
As previously described in #41150 ,
`InlineFormattingContext::font_metrics` contains info that are easily
accessible via `FontRef`, and the `FontInstanceKey` is already cached in
the `FontContext`. Instead of keeping an index and obtaining these info
using the index, one should be able to obtain the same info via a
`FontRef`.

Testing: This should not change the expected behavior and the current
unit tests and WPT testcases should be sufficient.
Fixes: #41150

---------

Signed-off-by: minghuaw <michael.wu1107@gmail.com>