servoshell: Do not forward key events to egui when no widget is focused (#41314)
When no egui widget is focused, the `WebView` itself is effectively
focused, so we should not forward keyboard events to egui. This change
does that as well as switch the key processing function in `Gui` to use
an early return approach for all events that should go immediately to
the `WebView` or are otherwise consumed.

Eventually, key events that are not consumed by the `WebView` should be
forwarded back to `egui`, so that the user can, for instance, tab out of
the `WebView`, but this requires having a version of winit with:
rust-windowing/winit#4026

Testing: This level of servoshell is untested unfortunately so no tests
are added.
Fixes: #41307.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>