storage: Introduce `ClientStorageProxy` as a helper for controlling the thread
This patch adds `ClientStorageProxy`, a helper on the frontend side for
communicating with `ClientStorageThread`. It abstracts over message types and
sending logic, making it easier to remotely control the thread.

The proxy currently only supports sending the Exit message, but it will be
extended to initiate new new storage key connections in the backend.

Signed-off-by: Jan Varga <jvarga@igalia.com>