script: Mark fetch-later requests as keep-alive
This implements the required plumbing for the keep-alive
flag on requests. The flag already exists on the request
object, but now also exists on the builder itself.

The flag is then passed into a canceller. While we could
make canceller optional, in many cases it isn't and would
require a whole lot more changes. To follow the spec more
closely, opted to put the keep_alive flag on the canceller
as well.

Note that this doesn't pass any extra fetch-later tests
since we are not correctly unloading iframe documents.
That will be fixed in a follow-up PR.

Part of #41230

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>