libservo: Make `UserInterfaceCommand` consistent and always forward mouse events to open egui dialogs (#41347)
This change addresses some inconsistencies in the way that events are
handled for `egui`. It moves event and `UserInterfaceCommand` handling
to
`ServoShellWindow` which allows sharing the implementation with the
embedded ports. In addition, moving event handling there makes it
simpler to always forward mouse events to egui dialogs when they are
open.

Fixes: #41339.
Testing: There is no real testing of this layer of servoshell, which is
why this regressed in the first place.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>