libservo: Add `SiteDataManager::list_sites` with storage type filtering
This patch adds a `list_sites` API to `SiteDataManager`, along with initial
infrastructure for specifying which storage types should be considered.

At this stage, sites correspond directly to storage origins. In the future,
this will be extended to support higher level grouping (for example by
eTLD+1).

This functionality is intended to provide a unified view of site data for
embedders.

Signed-off-by: Jan Varga <jvarga@igalia.com>