layout: Fix caret never rendering following a newline (#41330)
Fixes caret rendering when the caret is placed a the start of a line

```
Hello,

{here}World!
```

Also adds a utility function to the `range` crate.

Note that the rendering of the caret on many non-empty lines is still
broken due to the presence of several other bugs

Testing: I ran unit tests. Add a new reftest
Fixes: Part of #40839

---------

Signed-off-by: Freya Arbjerg <git@arbjerg.dev>