script: Implement customElementRegistry support for document.importNode
This method now allows you to pass in a custom registry.
The registry isn't used yet for callers, since we don't
support scoped registries. However, as we now pass in
the registry to `Node::clone`, we set the registry
when creating elements and cloning them.

As such, various tests start passing where we set the
registry. Some tests started failing, but they rely
on scoped registries which we don't support yet. These
tests thus flipped: those that were erroneously failing
are now passing and vice versa.

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>