script: Store `HashAlgorithmIdentifier` as `NormalizedAlgorithm`
For some dictionaries in SubtleCrypto interface, we store the hash field
in `HashAlgorithmIdentifier` type as a `SubtleKeyAlgorithm`. The
conversion to `SubtleKeyAlgorithm` is unnecessary. Moreover,
`SubtleKeyAlgorithm` is not supposed to be used there.

This patch fixes it by simply storing them as the normalized algorithm
given by Step 10.1.3 of normalization [1], for those hash field in
`HashAlgorithmIdentifier`.

[1] https://w3c.github.io/webcrypto/#algorithm-normalization-normalize-an-algorithm

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>