script: Implement quota computation for deferred fetches
For all upcoming and already-processed deferred fetches,
we now implement a quota. We don't support policy headers
yet to configure the quota, hence we use the default
quota. That's why we have some tests failing again.

It involves computing which document to compute the
quota for, as well as computing total length of requests.
Additionally, the DevTools computation for headers now
uses the same logic for headers as deferred fetches.

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>