script: Implement destroying of top-level navigable (#41355)
Another step towards being spec-compatible. Right now
we can't destroy an unsalvagable document during unloading,
as that breaks all kinds of tests. Before that, we need
to update various machinery to handle destroying correctly.

The first one is a top-level navigable, which explicitly
destroys a document after it is unloaded.

Part of #31973

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>