Add fling support for mouse wheel events
This add a WheelFlingHandler that manages fling state tied to
mouse wheel events. This is gated by the wheel_fling feature
that is not part of the default features: some platforms such
as MacOS manage their own fling when sending the MouseWheel
events.

Tested on Linux, this brings scrolling behavior on par with
Firefox and Chromium.

Signed-off-by: webbeef <me@webbeef.org>