layout: Stop clearing floats on anonymous blocks for intrinsic sizes
When computing the intrinsic sizes of a block container which doesn't
establish an inline formatting context, block-level children which don't
establish an independent formatting context clear floats. That's because
these blocks overlap the floats instead of being placed next to them.

However, we shouldn't do that if the block-level child is an anonymous
box that establishes an inline formatting context. This matches what
other browsers do.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>