base: Move Unicode offset types to `base`
This moves the Unicode offset types to the `base` crate and makes them
generally more usable throughout the Servo codebase. In addition, they
are renamed to use Rust naming and to be a bit more consistent:

 - `UTF8Bytes` -> `Utf8Length`
 - `UTF16CodeUnits` -> `Utf16Length`

There is also a bit more documentation explaining their use.

This is preparation for fixing issues with UTF-16 offsets in editable
text fields.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>