servoshell: ohos: Decouple initialization and window creation
We can start initializing servo without needing to wait for the
creation of a window.
This allows us to perform initalization of servo itself,
concurrently to the app UI thread setting up the app UI.
Once the native window is ready, we can then let servo create
the first window and load the initial URL.
This is also interesting in the context of using servo as a
webview library, where loading the library / initialising the
webview and loading the first url are typically decoupled
steps.

Follow-up PRs will add multi-window support, which means that some of
the multi-webview related code in this PR has open todos,
which will be addressed by follow-ups.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>