compositor: Remove unwrap if surface unbound (#41533)
If present_bound_surface fails for other reasons, we already handle this
by simply printing the error.
If there is no surface attached to the context, we also simply log the
error and otherwise ignore it,
instead of crashing servo.
Prevents a spurious crash when switching windows in #41532 (which is
probably also something that should be fixed, but the error should not
cause a crash in the first place).

Testing: Trivial, doesn't require extra testing.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>